Base Implementation of the Rush Hour game
In this notebook, there is base functions to simulate Rush Hour. Functions that will come in handy for the search algorithms, namely, a function to search for available moves on the grid, a function to find all the cars on the grid, and a function that moves a car and updates the grid.